+++ navauth-common/src/main/kotlin/pl/spcode/navauth/common/application/credentials/UserCredentialsService.kt
@@ -20,6 +20,11 @@ package pl.spcode.navauth.common.application.credentials
 
 import com.google.inject.Inject
 import com.google.inject.Singleton
+import java.util.UUID
+import java.util.concurrent.CompletableFuture
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import pl.spcode.navauth.common.application.credentials.queue.EncryptionQueueService
 import pl.spcode.navauth.common.domain.common.TransactionService
 import pl.spcode.navauth.common.domain.credentials.UserCredentials
 import pl.spcode.navauth.common.domain.credentials.UserCredentialsRepository
@@ -34,7 +39,10 @@ constructor(
   val credentialsRepository: UserCredentialsRepository,
   val transactionService: TransactionService,
   val credentialsHasherFactory: CredentialsHasherFactory,
+  val encryptionQueueService: EncryptionQueueService,
 ) {
+
+  private val logger: Logger = LoggerFactory.getLogger(javaClass)
 
   fun findCredentials(user: User): UserCredentials? {
     return credentialsRepository.findByUser(user)
@@ -72,14 +80,36 @@ constructor(
     credentialsRepository.deleteByUser(user)
   }
 
-  /** @param password the raw (not hashed) password */
-  fun verifyPassword(credentials: UserCredentials, password: String): Boolean {
-    require(credentials.hashedPassword != null) { "credentials do not have a password hash" }
+  /**
+   * @param password the raw (not hashed) password
+   * @throws
+   *   pl.spcode.navauth.common.application.credentials.queue.EncryptionTaskAlreadyQueuedException
+   *   if task is already queued
+   */
+  fun enqueueVerifyPassword(
+    credentials: UserCredentials,
+    password: String,
+    playerId: UUID,
+  ): CompletableFuture<Boolean> {
+    require(credentials.hashedPassword != null) { "credentials must have password hash" }
 
     val passwordHash = credentials.hashedPassword.passwordHash
     val hasher = credentialsHasherFactory.createHasher(credentials.hashedPassword.algo)
 
-    return hasher.verify(password, passwordHash)
+    val future = CompletableFuture<Boolean>()
+    encryptionQueueService.submitTask(playerId) {
+      try {
+        val result = hasher.verify(password, passwordHash)
+        future.complete(result)
+      } catch (ex: Exception) {
+        logger.error(
+          "Unexpected error occurred while trying to verify user id='${playerId}' password",
+          ex,
+        )
+        future.completeExceptionally(ex)
+      }
+    }
+    return future
   }
 
   fun hashPassword(password: String): HashedPassword {